ml/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm
google-genai>=1.2.0
matplotlib
faiss-cpu>=1.7.4
requests
duckduckgo-search>=5.0.0